% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oaTemplate.R
\name{blankPlot}
\alias{blankPlot}
\title{Create a Blank Plot}
\usage{
blankPlot(xlim, ylim)
}
\arguments{
\item{xlim}{x limits for the plot}

\item{ylim}{y limits for the plot}
}
\value{
none, plot is created on the current device
}
\description{
Create a Blank Plot
}
\author{
Jason Waddell
}

