% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/densityLegend.R
\name{colorPoly}
\alias{colorPoly}
\title{Function for plotting a colored polygon as part of a density legend}
\usage{
colorPoly(de1, tempDex, col, side)
}
\arguments{
\item{de1}{a density() object}

\item{tempDex}{a set of indices corresponding to the range of the current
segment to be plotted (which indices of the density object to ues)}

\item{col}{the color of the polygon to be plotted}

\item{side}{the side of the plot that the density legend should be plotted on}
}
\value{
none, graphics are added to the current device
}
\description{
Function for plotting a colored polygon as part of a density legend
}
\author{
Jason Waddell
}

