% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rounding.R
\name{customRound}
\alias{customRound}
\title{Custom rounding function to round to the nearest specified interval}
\usage{
customRound(x, roundTo)
}
\arguments{
\item{x}{numeric value(s)}

\item{roundTo}{rounding interval}
}
\value{
rounded numeric value(s)
}
\description{
Custom rounding function to round to the nearest specified interval
}
\author{
Jason Waddell
}

