% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findLocations.R
\name{findLocations}
\alias{findLocations}
\title{Returns a Vector of x Locations}
\usage{
findLocations(n, space, center)
}
\arguments{
\item{n}{number of observations for a given value}

\item{space}{space between points}

\item{center}{center plotting value}
}
\value{
numeric vector of location values
}
\description{
Returns a Vector of x Locations
}
\author{
Jason Waddell
}

