% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/densityLegend.R
\name{getBreaks}
\alias{getBreaks}
\title{Divide the range of x into intervals, returning the breakpoints of these
intervals}
\usage{
getBreaks(x, breaks, dig.lab = 3L)
}
\arguments{
\item{x}{a numeric vector which is to be converted to a factor by cutting}

\item{breaks}{a single number (greater than or equal to 2) giving the number
of intervals into which x is to be cut}

\item{dig.lab}{integer which is used when labels are not given. It determines
the number of digits used in formatting the break numbers}
}
\value{
a vector of numeric breakpoints
}
\description{
Divide the range of x into intervals, returning the breakpoints of these
intervals
}
\author{
Jason Waddell
}

