% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oaTheme.R
\name{oaTheme}
\alias{oaTheme}
\title{Apply OA ggplot2 theme}
\usage{
oaTheme(p, useOAColors = TRUE, expand = "both", bgColor = gray(0.9))
}
\arguments{
\item{p}{ggplot2 plot object}

\item{useOAColors}{boolean which indicates wether or not to use the oaColors
package to provide a color scheme. Default: TRUE}

\item{expand}{specify wether or not to expand the axis valid options are:
(both, x, y, none) Default: both}

\item{bgColor}{specify a different background color
(useful for plotting colors with alpha values) Default: gray(0.9)}
}
\value{
ggplot2 plot object
}
\description{
Apply OA ggplot2 theme
}
\author{
Willem Ligtenberg
}

