% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/densityLegend.R
\name{plotBars}
\alias{plotBars}
\title{A function for creating the segmented color bars in a density legend}
\usage{
plotBars(de1, side, colorPalette, colorBreaks)
}
\arguments{
\item{de1}{a density() object}

\item{side}{the side of the plot that the density legend should be plotted on}

\item{colorPalette}{A vector of color values}

\item{colorBreaks}{A vector of cutoff values for the color regions}
}
\value{
none, graphics are added to the current device
}
\description{
A function for creating the segmented color bars in a density legend
}
\author{
Jason Waddell
}

