% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/densityLegend.R
\name{plotDensityTrace}
\alias{plotDensityTrace}
\title{Function for plotting the density trace outline in a density legend}
\usage{
plotDensityTrace(de1, side)
}
\arguments{
\item{de1}{a density() object}

\item{side}{the side of the plot that the density legend should be plotted on}
}
\value{
none, graphics are added to the current device
}
\description{
Function for plotting the density trace outline in a density legend
}
\author{
Jason Waddell
}

