% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/densityLegend.R
\name{plotPolygonRegions}
\alias{plotPolygonRegions}
\title{Function to plot all colored density regions of a density legend}
\usage{
plotPolygonRegions(de1, side, colorPalette, colorBreaks)
}
\arguments{
\item{de1}{a density() object}

\item{side}{the side of the plot that the density legend should be plotted on}

\item{colorPalette}{a vector of color values}

\item{colorBreaks}{a vector of cutoff values for the color regions}
}
\value{
none, graphics are added to the current device
}
\description{
Function to plot all colored density regions of a density legend
}
\author{
Jason Waddell
}

