% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pointsOnBoxplot.R
\name{pointsOnBoxplot.formula}
\alias{pointsOnBoxplot.formula}
\title{Draw Points on Top of a Boxplot using Appropriate Shifting}
\usage{
\method{pointsOnBoxplot}{formula}(formula, data = NULL, ...,
  na.action = NULL)
}
\arguments{
\item{formula}{a formula of the form a ~ b (+ c, etc.), where a is a numeric vector and all other variables are categorical}

\item{data}{an optional input parameter of a data.frame containing the variables used in the formula}

\item{...}{further arguments to be passed to pointsOnBoxplot.default}

\item{na.action}{parameter specifying how to handle missingness}
}
\description{
Draw Points on Top of a Boxplot using Appropriate Shifting
}
\author{
Jason Waddell
}

