% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/densityLegend.R
\name{splitColorVar}
\alias{splitColorVar}
\title{Function to take a numeric vector 'colorVar' and palette 'colorPalette', and
return a list containing a vector of color assignments for each element of
'colorVar' (to be used in plot calls), and a vector of breaks defining the
color regions (to be used in densityLegend)}
\usage{
splitColorVar(colorVar, colorPalette, breaks = NULL)
}
\arguments{
\item{colorVar}{the numeric vector of values used to color the points}

\item{colorPalette}{a color palette. If 'colorPalette' contains, for example,
6 colors, then the values of colorVar will be split and assigned to these 6
colors}

\item{breaks}{(optional) a numeric vector of two or more unique cut points}
}
\value{
a list containing a vector of color assignments ('colorVec') for each
element of 'colorVar' (to be used in plot calls), and a vector of breaks
('breaks') defining the color regions (to be used in densityLegend)
}
\description{
Function to take a numeric vector 'colorVar' and palette 'colorPalette', and
return a list containing a vector of color assignments for each element of
'colorVar' (to be used in plot calls), and a vector of breaks defining the
color regions (to be used in densityLegend)
}
\author{
Jason Waddell
}

